BinarySearchTreeChecker
this is an assignment from my data structures class done on repl.it where I was given  a binary tree represented as an array.  I had to verify if the input was a binary search tree or not. the Inputs were given as a space-separated arrays representing a binary, with the root node occupying the 0th index, the root node’s children occupying indices 1 and 2, their children occupying indices 3-6, etc.  All trees will be balanced, and will be filled left to right in the bottom layer.  


Thus, if given the following input:

10 5 15 2 7 11 25 1
then it will print:
true since this tree is indeed a binary search tree. As another example, the
input
2 4 5
does not represent a binary search tree since the number 2 is less than both its children (4 and 5).

it runs in O(n) time

Code for reading input and writing output was provided in the assignment my  job was to complete the isBinarySearchTree() function while maintaining O(n) time without using any additional libraries.