Create BasicCalulator
This is an assignment from my data Structures class this code evaluates expressions that contain numbers, +, *, (, and ).  it takes in a string representing an expression and returns an integer representing its result. Assuming the input string will not contain any spaces.  The following are examples of input and corresponding output:

Input:
3+5

Output:
8

Input:
3+5*7

Output:
38

Input:
3*(5*(7+2))

Output:
135

The functionality for reading and printing answers was provided but I had to  complete the eval() method  which  takes in a string representing an arithmetic expression as described above and outputs a single integer representing the result of this expression.